requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
et_xmlfile==2.0.0
execnet==2.1.1
Faker==37.3.0
idna==3.10
iniconfig==2.1.0
javaobj-py3==0.4.4
jsonpath==0.82.2
jsonpath-ng==1.7.0
openpyxl==3.1.5
packaging==25.0
pluggy==1.5.0
ply==3.11
pycryptodome==3.23.0
PyMySQL==1.1.1
pytest==8.3.5
pytest-assume==2.4.3
pytest-dependency==0.6.0
pytest-ordering==0.6
pytest-repeat==0.9.4
pytest-rerunfailures==15.1
pytest-xdist==3.6.1
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
setuptools==80.4.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
